// models/InterestsResponseDto.ts
/**
 * Bookish API
 * The bookish API description
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { InterestsPayloadDto } from '../models/InterestsPayloadDto';
import { HttpFile } from '../http/http';

export class InterestsResponseDto {
    'data': Array<InterestsPayloadDto>;
    'isSuccess': boolean;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "data",
            "baseName": "data",
            "type": "Array<InterestsPayloadDto>",
            "format": ""
        },
        {
            "name": "isSuccess",
            "baseName": "isSuccess",
            "type": "boolean",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return InterestsResponseDto.attributeTypeMap;
    }

    public constructor() {
    }
}